Move BaseDbtPackageTests to shared tests/packages/ module
